Update shop.py
Shop 2.0
=====================================================
- Added Multi-Server support. Each server will have a different shop list
- Items in the shop can have quantities or set to infinite. If the quantity runs out the item is removed
- Can set a discount for items in the shop.
- Trash an item you no longer want to have. (Destroys all instances of the item)
- Can set for manual purchased items or automatic. See Automatic Replies for more information
- Check your shop version
- Pending list complete overhaul
- Display 12 pending items at a time, seperated by pages. No limit to pending items!
- Search for a user's pending items or by a confirmation code
- Users can look up their confirmation code or can check on all their pending items
- Status can be set for pending items.
- Setting a notify role will now check if the role exists on the server
- Admins can give an item to a user

Automatic Shop Function
You can setup automatic shop replies when a user purchases an item. To change to automatic shop, use the command:
<p>setshop ptype
You will be asked to agree to the warning message before you can proceed.
Add an item to the shop like normal, and set a quantity. 
Then you can set multiple buy messages to the item, using the command:
<p>setshop buymsg <itemname>
If you set 5 buymsgs then it will pick one random buy msg with no repeats, and send that to the user via Private Message.
Make sure you close down the shop, using he command:
<p>setshop toggle
So that users do not purchase the item before you can set a buy message. In the event that they do purchase an item, and there is no buy message, they will be prompted that they need to contact an Admin.
You can set an item to infinite and add buy messages until they run out, and get prompted with this error. 
Please know that this not a secure method of using shop, because I cannot protect against all methods of credit manipulation from other cogs, and that I am not responsible for any loss of information using this method.