Casino Patch 1.6.94
Converted cool down time from seconds into Unix time for better tracking. This also aligns us with economy.
You will need to get the python-dateutil library for this update (SORRY!).
Fixed an issue where cool downs would not display properly in stats.
Fixed an issue where the membership updater would not work for "days on server" requirements.
Fixed a logging error in blackjack and war.
Fixed an issue when creating a new membership and the requirement did not display correctly.
Fixed an issue when transferring legacy users to the new casino.